post working
kind of questionable if this is how it should work. Just passing a string in the request body and decoding on the server feels weird, but maybe it's right. I'm surprised I don't need to pass it an entire object. I'll look into this more as I'm working on put.